Move tests into profile package
