+++ src/test/java/seedu/address/logic/parser/profile/AddProfileCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.profile;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
@@ -28,7 +28,6 @@ import static seedu.address.testutil.TypicalProfiles.BOB;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.profile.AddProfileCommand;
-import seedu.address.logic.parser.profile.AddProfileCommandParser;
 import seedu.address.model.profile.Email;
 import seedu.address.model.profile.Name;
 import seedu.address.model.profile.Phone;
